prettify avatar

+++ modules/Utility/color/avatar.js
@@ -1,19 +1,23 @@
-const vibrant = require('node-vibrant');
-const request = require('request-promise-native');
-const Discord = require('discord.js');
+const vibrant = require("node-vibrant");
+const request = require("request-promise-native");
+const Discord = require("discord.js");
 
-const generateImage = require('./image/image.js');
-const assign = require('./assign.js');
+const generateImage = require("./image/image.js");
+const assign = require("./assign.js");
 
 module.exports = async function(msg, menusOpen) {
-
     // User avatar url in png format
-    const url = msg.author.displayAvatarURL.substr(0, msg.author.displayAvatarURL.lastIndexOf('.'));
+    const url = msg.author.displayAvatarURL.substr(
+        0,
+        msg.author.displayAvatarURL.lastIndexOf(".")
+    );
 
-    let img = await request.get({url: url, encoding: 'binary'});
+    let img = await request.get({ url: url, encoding: "binary" });
 
     // Generate swatches from image buffer
-    const swatches = await vibrant.from(Buffer.from(img, 'binary')).getSwatches();
+    const swatches = await vibrant
+        .from(Buffer.from(img, "binary"))
+        .getSwatches();
 
     // flush img from memory
     img = null;
@@ -21,41 +25,51 @@ module.exports = async function(msg, menusOpen) {
     // Go through swatches and select non-null colors
     let i = 0;
     let choices = [];
-    choiceString = '';
+    choiceString = "";
     for (let key in swatches) {
-        if (swatches[key] == null)
-            continue;
+        if (swatches[key] == null) continue;
 
         choices.push(swatches[key].getHex());
         choiceString += `${i + 1} -> ${swatches[key].getHex().toUpperCase()}\n`;
 
         i++;
-    }   
+    }
 
     const choiceImg = await generateImage(choices, msg);
-    const choiceMsg = await msg.channel.send('Select a color\n"exit" to exit the menu', { files: [choiceImg] });
+    const choiceMsg = await msg.channel.send(
+        'Select a color\n"exit" to exit the menu',
+        { files: [choiceImg] }
+    );
 
-    const msgC = new Discord.MessageCollector(msg.channel, a => a.author == msg.author, { time: 300000 });
-    msgC.on('collect', async message => {
-
+    const msgC = new Discord.MessageCollector(
+        msg.channel,
+        (a) => a.author == msg.author,
+        { time: 300000 }
+    );
+    msgC.on("collect", async (message) => {
         // Allow exiting out of the menu
-        if (message.content.toLowerCase() == 'exit') {
+        if (message.content.toLowerCase() == "exit") {
             msgC.stop();
             return;
         }
 
         if (!isNaN(message.content)) {
             try {
-                if (Number(message.content) > 0 && Number(message.content) <= choices.length) {
+                if (
+                    Number(message.content) > 0 &&
+                    Number(message.content) <= choices.length
+                ) {
                     const cr = choices[Number(message.content) - 1];
                     await assign(msg, cr.toUpperCase());
 
                     msgC.stop();
                     return;
                 } else {
-                    await msg.channel.send(new Discord.RichEmbed()
-                    .setColor('RED')
-                    .setDescription('That\'s not a valid choice, is it'))
+                    await msg.channel.send(
+                        new Discord.RichEmbed()
+                            .setColor("RED")
+                            .setDescription("That's not a valid choice, is it")
+                    );
                 }
             } catch (err) {
                 throw err;
@@ -63,9 +77,8 @@ module.exports = async function(msg, menusOpen) {
         }
     });
 
-    msgC.on('end', () => {
-        if (choiceMsg.deletable)
-            choiceMsg.delete();
+    msgC.on("end", () => {
+        if (choiceMsg.deletable) choiceMsg.delete();
 
         menusOpen.splice(menusOpen.indexOf(msg.author.id), 1);
     });
